Fix rare potential for hang in NO_IPCRM crash tests (e.g. manually_start/dual_fail2_no_ipcrm1 subtest) due to DBDANGER and instance freeze
We saw a hang in the manually_start/dual_fail2_no_ipcrm1 subtest in one run across many runs.
Turns out in this case, a DBDANGER showed up in the syslog (because of the kill -9 happening
at a wrong window while inside the transaction commit logic) and that caused an instance freeze
which was never resolved resulting in a test hang. The fix is to remove DBDANGER from the
custom errors file like is already done in mupip/u_inref/mu_rundown_no_ipcrm1.csh.

That code is now moved to a new script com/adjust_custom_errors_for_no_ipcrm_test.csh which
is now invoked from various tests that crash running mumps processes and have "NO_IPCRM" (i.e.
do not remove database shared memory) leaving the possibility of a DBDANGER and instance freeze
(if custom errors scheme is randomly defined in the test).