[#361] New r124/ydb361 subtest (tests https://gitlab.com/YottaDB/DB/YDB/issues/361)
This test fails approximately 1 out of 10 test runs on the ARMV7L box with the assert
that #361 describes. In all failures till now, the test system env var "gtm_test_tls"
has been always TRUE. So the slowness of TLS operations (which happen just before adding
a new history record) on the ARMV7L contributes to making this test fail reasonably
reliably there. This test does not fail on the x86_64 boxes currently.

Given this test fails well on ARMV7L without the #361 code fixes, this is considered good
enough of an automated testcase.