Fix rare online_bkup/online4 subtest failure due to source server writing to the journal file header when test was not expecting any one to update the journal file
This test wants to ensure the live journal file is not touched by the SET ^a command on the backup db.
And the way it checks this is by comparing the timestamp of acnm.mjl before and after the SET ^a. And
in a failure we saw, they happen to be different even at the minute level (that ls -l captures). Below
is the relevant portion of the diff file.

> cat online4.diff
43c43,46
< No difference seen
---
> 5c5
> < -rw-rw-rw- 1 user grp   69632 Jun 28 03:07 acnm.mjl
> ---
> > -rw-rw-rw- 1 user grp   69632 Jun 28 03:08 acnm.mjl

This is a case where the file size is the same but timestamp is different. Taking a look at the journal
file header after the SET ^a command (in the "after_back_update" subdirectory), I see the following in
the acnm.mjl file header.

> mupip journal -show=header -forward -noverify after_back_update/acnm.mjl
.
.
Process That Created the Journal File:

PID        NODE             USER         TERM JPV_TIME
-----------------------------------------------------------------
0000001991 node12           user              2018/06/28 03:07:53

Process That First Opened the Journal File:

PID        NODE             USER         TERM JPV_TIME
-----------------------------------------------------------------
0000001539 node12           user              2018/06/28 03:08:40

Pid 1991 is the mupip backup process as that is the one which created the journal file (backup switches
journal files). This happened at 3:07:53. But one can see another pid 1539 that opened the journal file
at 03:08:40. Almost a minute later. That is the cause of the test failure.

Some more searching showed this is the source server.

> cat SRC_03_06_11.log
Thu Jun 28 03:06:14 2018 : mupip replic -source -start -secondary=127.0.0.1:36010 -buffsize=33554432 -instsecondary=INSTANCE2 -log=online_bkup_1_1/online4/SRC_03_06_11.log -tlsid=INSTANCE1 -jnlfileonly
Thu Jun 28 03:06:14 2018 : %YDB-I-REPLINFO, GTM Replication Source Server with Pid [1539] started for Secondary Instance [INSTANCE2]

It had opened the journal file in shared memory as part of catching up with the backlog (and opening a
journal file causes one to write to the jnl file header on disk even if it is the source server) which
caused the test failure.

In this case, the source server was reading from journal files because the test had forced it.

> cat settings.csh
92 # JNLFILEONLY enabled by do_random_settings.csh
93 setenv gtm_test_jnlfileonly 1

Changing the test to force gtm_test_jnlfileonly to 0 can avoid these kinds of rare failures.

Additionally, the test time capturing scheme has been enhanced to use ls -l --full-time (instead of ls -l)
as that will give us micro-second level granularity. Right now we are at a minute level granularity and
so even if an update happens within the same minute, it will go unnoticed whereas we expect no update
to happen even at the micro-second granularity.