Fix rare r124/ydb282srcsrvrerr failure (and potential v63004/gtm8922 subtest failure) by setting flush timer to 1 hour (not 1 minute)
A prior commit addressed the r124/ydb282srcsrvrerr subtest failure by setting the flush timer
to 1 hour. For that it specified a timer of 1:0:0. This turns out to be 1 minute. 1 hour is
represented as 1:0:0:0. In an internal test run we had a test failure because a relatively slow
host took more than 1 minute to finish a section of this test which meant the flush timer kicked
in and caused a hang like seen before. This is now fixed by setting the flush timer to 1 hour
for real (1:0:0:0).

A similar issue in the v63004/gtm8922 subtest driver script is also fixed.