Fix v63003/gtm8182 subtest failure on systems where test output directory did not have an <extra> in the fullpath
The v63003/gtm8182 subtest failed on a new host where the test output was redirected to a fullpath
that did not have the string "extra" in it when the test script was expecting that. This seemed
like an unnecessary expectation so the grep has been changed to search for mumps.repl from the
GDE SHOW -INSTANCE output to just filter out the line containing the replication instance file.

While at this, enhanced the test so multiple show outputs on each instance go to different output files.