Fix rare v60000/gtm6571 test failure due to process taking more than 30 seconds to die on a loaded system
We saw a failure in internal testing with the following diff.

> JOB 33236 did not die in 30 seconds
> JOB 33242 did not die in 30 seconds

It is not clear exactly how long the process took to die but since that system was heavily loaded at
that point in time (was running 4 tests concurrently), 30 seconds might not be enough in some cases.
So the timeout is bumped to 300 seconds (5 minutes).