Fix rare v60000/dbioerr subtest failure due to DBIOERR message in syslog from a concurrent test
We had a test failure with the following diff

> Oct 03 16:51:51 hostname YDB-MUMPS[820042]: %YDB-I-DBIOERR, Error while doing write operation on region RGI (UKplk) --

This meant the test saw the DBIOERR message it was expecting and saw one more message in the syslog.
That was from a concurrently running test. To fix this issue, we search for DBIOERR syslog messages
which also have the current test output directory in them.