Fix rare stress/concurr_replwason failure due to slow system taking longer than 5 minutes to clear receiver backlog
We had a failure of this test on an ARMV6 box (Pi Zero) with the following output

==Executing MULTISITE_REPLIC 'RUN INST2 com/close_reorg.csh >>& stress_reorg.out'==
TEST-E-MULTISITE replic action failed.Pls. check msr_execute_27.csh and all logs related to timestamp 06_01_19_27
=>Last seqno received at INST3 is GTM_TEST_DEBUGINFO 36157

The corresponding test script invocation is the following.

$MSR RUN INST3 "set msr_dont_trace ; $gtm_tst/com/wait_until_rcvr_backlog_clear.csh"

The wait_until_rcvr_backlog_clear.csh checks the backlog every 5 seconds for 60 iterations in what
should take it 300 seconds (5 minutes). But in the actual test run, the 60 iterations took 10 minutes
instead. And in this time, the backlog had reduced from 16941 to 9074. But since it did not become zero,
the script exited with a non-zero status causing the test failure.

> cat rb_1968_060142.log
16941 : number of backlog transactions received by receiver server and yet to be processed by update process
44771 : sequence number of last transaction received from Source Server and written to receive pool
27830 : sequence number of last transaction processed by update process

> cat rb_4032_061154.log
9074 : number of backlog transactions received by receiver server and yet to be processed by update process
44771 : sequence number of last transaction received from Source Server and written to receive pool
35697 : sequence number of last transaction processed by update process

Given 400 more seconds or so the script would have run fine. So the timeout is bumped up from 300 seconds
to 1800 seconds which should be ample enough even for a slow system. This is the same timeout that a
similar script com/wait_until_src_backlog_below.csh already uses.