[#362] Fix suppl_inst_B test failure in multi-host mode now that supplementary_err subtest runs always
Recently the suppl_inst_B test was enhanced to run the supplementary_err subtest always. But
the reference file was only partially fixed. It was still suspending this subtest in the
MULTISITE mode (i.e. multi-host test run). But instream.csh was running this subtest even in
that condition. This is now fixed by removing all SUSPEND_OUTPUT/ALLOW_OUTPUT macros around
the supplementary_err subtest in outref.txt.