Fix rare v62000/gtm8086 test hang in recent JNLSWITCHRETRY test changes (https://gitlab.com/YottaDB/DB/YDB/issues/235)
We had a test hang in-house where the test was hung in line 99 below.

v62000/u_inref/gtm8086.csh
     97         if ($jnlswitchretry) then
     98                 $MSR STARTRCV INST1 INST2 >& jnlswitchretry_2.log
     99                 $MSR SYNC INST1 INST2 >& jnlswitchretry_3.log
    100         endif

The backlog was not clearing because the source server was stuck waiting for crit with the following C-stack.

(gdb) where
 #4  mutex_long_sleep (addr=0x7f90ceadd000, csa=0x68e040, mutex_spin_parms=0x7f90cfefd054) at sr_unix/mutex.c:403
 #5  gtm_mutex_lock (reg=0x7285f0, mutex_spin_parms=0x7f90cfefd054, crash_count=0, mutex_lock_type=MUTEX_LOCK_WRITE) at sr_unix/mutex.c:856
 #6  grab_crit (reg=0x7285f0) at sr_unix/grab_crit.c:114
 #7  repl_ctl_create (ctl=0x7ffd3ffe9e20, reg=0x7285f0, jnl_fn_len=0, jnl_fn=0x0, init=1) at sr_port/gtmsource_ctl_init.c:181
 #8  gtmsource_ctl_init () at sr_port/gtmsource_ctl_init.c:436
 #9  gtmsource_get_jnlrecs (buff=0x7b33f8 "\001", data_len=0x7ffd3ffe9f4c, maxbufflen=2097144, read_multiple=1) at sr_unix/gtmsource_process_ops.c:976
 #10 gtmsource_process () at sr_unix/gtmsource_process.c:1523
 #11 gtmsource () at sr_unix/gtmsource.c:528

It was waiting for crit held by another process which in turn was stuck waiting for the instance freeze
to clear with the following C-stack.

 #7  0x00007f92d692865c in wait_for_repl_inst_unfreeze_nocsa_jpl (jpl=0x1ad7e40) at sr_port/anticipatory_freeze.h:490
 #8  0x00007f92d6928a0a in wait_for_repl_inst_unfreeze (csa=0x1b3e040) at sr_port/anticipatory_freeze.h:513
 #9  0x00007f92d692a020 in jnl_write_attempt (jpc=0x1b3fe40, threshold=3102928) at sr_port/jnl_write_attempt.c:335
 #10 0x00007f92d68508e2 in jnl_flush (reg=0x1ac92d0) at sr_port/jnl_flush.c:57
 #11 0x00007f92d697bbe9 in tp_tend () at sr_port/tp_tend.c:762
 #12 0x00007f92d6814a8c in op_tcommit () at sr_port/op_tcommit.c:496
 #13 0x00007f92d69c5a37 in gvcst_put2 (val=0x1d32370, parms=0x7ffc4e28d7a0) at sr_port/gvcst_put.c:2875
 #14 0x00007f92d69ad467 in gvcst_put (val=0x1d32370) at sr_port/gvcst_put.c:299
 #15 0x00007f92d6900dee in op_gvput (var=0x1d32370) at sr_port/op_gvput.c:74

This is a case where a process in TP commit logic realizes it is time to extend the journal file and so
requires crit. This is now worked around by bumping the journal file allocation/extension to be the
same as the autoswitchlimit (i.e. 16384) that way one never needs to invoke jnl_flush() while inside
tp_tend() or t_end(). This way the source server will not be stuck eternally waiting for crit while
it wants to open the journal file in gtmsource_ctl_init().

That said, this is just a specific scenario observed in this test hang. It is possible a few other such
hangs can happen where a mumps process holds crit while waiting for the instance freeze to be turned off.
Each of those cases needs to be addressed on a case-by-case basis since we want to test the JNLSWITCHRETRY
logic for which we do need the source server to reconnect to the receiver server while the instance is
frozen.

While analyzing this failure, a typo in com/do_random_settings.csh was noticed and it is now fixed.