[#362] Fix rare manually_start/dual_fail2_no_ipcrm1 and manually_start/dual_fail3_nonsym subtest failure with dbcheck extract differences between pri and sec side
We had an in-house test failure where the manually_start/dual_fail2_no_ipcrm1 subtest had a db
extract difference between pri and sec side. This turns out to be the exact same issue that we saw
in the manually_start/dual_fail2_no_ipcrm2 subtest which was fixed in a prior #362 commit.

Like there, the fix is to do a fetchresync rollback on the secondary after a crash. Previously the test
used to do a mupip rundown -override followed by a mupip_rollback.csh invocation which in turn did a
mupip rundown -override. Since we only want the rollback without the rundown -override invocations,
all of those are removed and replaced with a specific mupip journal -rollback -fetchresync command.

While at this, a similar potential issue in manually_start/dual_fail3_nonsym subtest is also fixed.