[#361] Enhance r124/ydb361 subtest to test GTMASSERT2 error in update process
See https://gitlab.com/YottaDB/DB/YDB/merge_requests/415 for more details on the original
code issue. This was not exercised by the r124/ydb361 subtest because it did not do any
updates. That is now fixed by doing 10 updates and ensuring they get replicated fine.

This test fails once in 10 runs or so on the ARMV7L boxes with the GTMASSERT2 when run
without the above mentioned merge request fix.