Fix rare v62002/gtm8317 subtest failure due to delayed INVTMPDIR message from gtmsecshr in syslog
We had a failure in a Pi Zero box (ARMV6L) with the following diff

9c9
< 1
---
> 0

When gtm_tmp env var is set to a bogus value, we expect INVTMPDIR from gtmsecshr but did not
find any in this particular test run. This turned out to be a test issue. The message did show
up in the syslog eventually but did not at the time the getoper.csh call happened.

And that is because gtmsecshr runs in the background and on a slow system it could take a while
for the INVTMPDIR syslog message to show up in the journalctl logs. Therefore the getoper.csh
call should also provide "INVTMPDIR" as an additional parameter to indicate wait until this search
string is found in the syslog.

While at this, other calls to getoper.csh where there is a mumps invocation (which is not a
backgrounded process like gtmsecshr) have been enhanced to not do unnecessary sleeps and not
pass an end time but instead pass "" as the after-time ($2) which getoper.csh knows to handle.

There is just one case where a sleep 1 is not removed and that is the last test where we do not
expect any messages to show up from gtmsecshr. In that case, there is no specific message to wait
for so we sleep a little extra just in case some INVLINKTMPDIR message we do not expect shows up.