Fix v62002/gtm6638 test failure by bumping timeout from 10 or 15 minutes to 1 hour
We recently changed the filesystem used for test outputs from ext4 to xfs on a box that had 4 CPUs.
After the change, the v62002/gtm6638 subtest started failing with different symptoms based on
whether simpleapi was randomly chosen or not.

If threeen1f.m was chosen (i.e. no simpleapi), the timeout was set to 10 minutes but the test took
more time than that and failed.

If simpleapi_threeen1f.c was chosen (i.e. simpleapi), the timeout was set to 15 minutes but the
test took more time than that and failed.

To avoid such issues, the timeout has been bumped to 1 hour in both cases. The test passes reliably
now on this box after the timeout change (it now takes approximately 25 minutes to run).