Fix rare r124/ydb358 subtest failure due to EINTR in waitpid() (possible due to db flush timer)
We had a failure in internal testing where an assert in line 59 below failed.

r124/inref/ydb358.c
-------------------
     56   /* Wait for child to terminate */
     57   printf("Parent pid : Waiting for child to terminate\n"); fflush(stdout);
     58   ret = waitpid(child, &stat, 0);
 --> 59   YDB_ASSERT(-1 != ret);

We have a core file but we need the "errno" from the core file which is not an easy task.
But I suspect the errno was EINTR due to a db flush timer (the parent does a ydb_set_s()).
And waitpid() can return -1 due to EINTR.

This should hopefully fix the failure. Even otherwise, have enhanced the C program to save
the errno in a local variable so we get the real errno next time such a failure happens.