Fix rare simpleapi/locks subtest failure (and similar potential issue in stress/concurr subtest) due to EINTR in waitpid() (possible due to db flush timer)
We got a failure in simpleapi/locks test where the return from a waitpid() call was -1.
The suspicion is that waitpid() failed due to an EINTR because the parent did a database set
which in turn started a flush timer that popped in the middle of the waitpid().
The fix is to add a do/while loop to check for EINTR and restart the waitpid() call.

While at this, a similar potential issue in the stress/concurr subtest is also fixed
(module stress/inref/simpleapi_run_concurr.c).

This fix is similar to commit SHA 733968e26fec6fdd9625422a41ca4b54801b702f, the r124/ydb358 failure
fix done a day before.