Fix rare v60000/anticipatory_freeze_utilities subtest hang due to foreground unfreeze finishing before backgrounded freeze
We had a test hang in in-house testing where the backgrounded MUPIP FREEZE -ON had only frozen AREG
when the MUPIP FREEZE -OFF command was issued in the foreground. The unfreeze command then only
cleared the freeze in AREG. After that, the MUPIP FREEZE -ON command froze DEFAULT in the background
and this caused the next stage of the test (MUPIP SET JOURNAL) to hang because of a frozen region.