Fix rare v54003/C9K08003315 subtest failure due to a test timing issue
We had a failure of this test in a loaded system with the following line missing in the actual output
of the mupip journal recover -nochecktn which was interrupted by a Ctrl-D.

%YDB-I-MUJNLSTAT, End processing at ... ... .. ..:..:.. 20..

This is suspected to be because the wait for the spawned process to finish was done using a hardcoded
timeout (expect -timeout 30). If the process was still running after 30 seconds (due to a loaded/slow
system), we would then go on to spawn a new mupip journal recover process (the one where N is entered
explicitly). And once the second journal recover process is done, we would exit the expect script
which would kill all the spawned off processes. I suspect the first recovery process was abruptly
terminated (before it had a chance to terminate gracefully).

Therefore the expect script is now reworked to wait for the spawned off child to finish (expect eof)
before moving on to the next spawn.

While at this, added a few comments in the reference file to make it more readable as to what is
being tested.