Fix rare v62000/zsyslog subtest failure in loaded system
We had a rare test failure with the following diff

4,5d3
< ##DIFFERENT EXPRESSION##^.* YDB-MUMPS\[[0-9]*\]\:.* Very bogus syslog message \#1 from v62000\/zsyslog test that is sort-of long$## VS ####
< ##DIFFERENT EXPRESSION##^.* YDB-MUMPS\[[0-9]*\]\:.* Another shorter bogus syslog message \#2 from v62000\/zsyslog test$## VS ####

The bogus messages which were expected did not show up in the syslog. This is a test issue
where it waits 3 seconds and extracts all messages until then using getoper.csh whereas it
should instead use "" as the after-time and getoper.csh has logic to correctly extract all
messages as of then by sending a random syslog message and waiting for it to show up.