Fix occasional v54003/C9K08003315 subtest failure due to a test timing issue
A recent change to this subtest (to fix another test timing issue) introduced another timing issue
which showed up only once in a while in a particular Ubuntu 18.04 host.

The issue is that we were waiting for "Proceed? [Y/N]: " to show up but we should have also waited
for the accompanying newline "\r" to show up along with it or else it was possible for the following
puts to happen in between the expected string and the "\r" causing a different output (and in turn
a test failure). This is now fixed by adding "\r" to all expect commands that were recently added.