Fix occasional r124/ydb313 subtest failure when lots of ipcs are left over
The r124/ydb313 subtest only failed when there were lots of ipcs left over (which happens in a
nightly testing) whereas it never failed when all ipcs were cleaned.

The diff had lines like the below.

< jnlpool Semaphore ID found
< recvpool Semaphore ID found

The issue is that "ipcs -s | grep -w -q $jnlSem" exited with an exit status of 141, a non-zero value
which in turn caused the expected message to not be printed and caused the test failure.

grep -q exits with a zero status immediately as soon as a match is found. But since there are lots of
semaphores lying on the system, ipcs -s is still writing to the pipe and since there is no reader (the
grep process has exited), the ipcs process is sent a SIGPIPE signal from the OS and it exits with
a status of 141.

When there are no ipcs left around on the system, the ipcs -s process is most likely to not be writing
to the pipe after the grep has exited. And so the failure does not happen in that case.

The fix for this occasional failure is to avoid using the -q in a shell pipeline. Instead the output/error
is redirected to /dev/null that way it acts as the -q but without the SIGPIPE signal. With this fix, the
test passes reliably even if a lot of ipcs are left around in the system.