Fix rare v63003/gtm8787 subtest failure due to KILLBYSIG error from unrelated/concurrently-running subtest
The v63003/gtm8787 subtest failed in in-house testing with the following diff.

3a4
> %YDB-F-KILLBYSIGUINFO, YottaDB process 94007 has been killed

The test did not expect to see a KILLBYSIGUINFO message from the MUPIP JOURNAL EXTRACT process it
spawned off in the background but it did see one in the failed run. This one turned out to be a
message from an unrelated process from some other concurrently running subtest.

The gtm8787 subtest is now enhanced to search for messages in syslog from the journal extract process
by recording the pid of the backgrounded process and using that in the grep. This should avoid
concurrent subtests from affecting this subtest.