[#376] Fix 64bittn/rolling_upgrade subtest failure due to older versions not supporting TLS 1.3 (https://gitlab.com/YottaDB/DB/YDB/issues/376)
If the test framework randomly enables TLS and the current system has TLS 1.3 (OpenSSL 1.1.1)
(e.g. if it is a Ubuntu 18.10 box), then pre-r1.24 source/receiver server will assert fail if they
talk to a r1.24 source/receiver server in a replicated environment since they do not support TLS 1.3.
Therefore disable TLS in the test if an older version that does not understand TLS 1.3 is randomly chosen.