[AARCH64] Fix v62000/gtm8047 subtest failure on 64-bit ARM
strace -k shows up as a supported feature on 64-bit ARM in a "strace -h" output
but when used it gives the following error.

strace: walk: cannot initialize libunwind: Input/output error

So the test has been enhanced to not rely on "strace -h" but instead run "strace -k"
for a simple case and check if $status is 0 and if so use it with "mumps -run".