Enhance v55000/GTM6813 subtest to reduce runtimes and display more helpful information in case of a failure
This test currently fails once in a while on the 64-bit ARM box. It calls a function adder() with
5 parameters that are related to each other (e.g. the 5th parameter i4 should be 5 times the first
parameter i0 etc.). This adder() call is done in a loop until 5000 interrupts (MUPIP INTRPT) are seen
(which are externally sent by a jobbed off process). Once in a while, we see the passed in parameter
i4 is not exactly 5 times the first parameter i0. This is likely a code issue in the 64-bit ARM port.

But to help with analyzing test failures, the test now records the difference whenever it notices
an incorrect call (instead of finally reporting a sum that is incorrect). This way we have more
information on exactly how many calls to adder() had issues and by how much the difference was in
each call. The local variable "error" maintains this information if ever there is a discrepancy.

This test took nearly 2 hours to run on the 64-bit ARM box. That is considered too long so the test
is enhanced to run until 100 interrupts are seen (should reduce the test runtime 50 times without
affecting test coverage) AND the for loop has run for at least 30 seconds.