Fix rare suppl_inst_B/supplementary_err subtest failure when test framework randomly chooses multi-host mode
We had a failure of this subtest with the following diff.

> REMOTEHOST:suppl_inst_B_1_2/supplementary_err/instance4/START_19_27_19_35.out
> %YDB-E-SRCSRVNOTEXIST, Source server for secondary instance INSTANCE1 is not alive

The actual failure was in a remote host where the source server startup ($MSR STARTSRC) failed
with a SRCSRVNOTEXIST error. The test sets env var gtm_test_repl_skipsrcchkhlth to 1 to specifically
bypass this situation. But in a test that runs across multiple hosts, a setenv in a local script does
not carry the env var to the remote side where the $MSR command actually runs. The fix is to do
the setenv in a file called env_supplementary.csh on the remote instance BEFORE the actual $MSR STARTSRC
or $MSR STARTRCV command.

Although the failure occurred in one stage of the test, various other stages of the test had similar
issues and all of them are now fixed.