Fix rare socket/socbasic test failure when tls is enabled and TLS1.3 is used
This subtest used to fail once in a few test runs with the following diff.

< 8 starting ... finished PASS
---
> 8 starting ... finished TEST-E-FAILED x
> ^dataasalongvariablename45678901(8,"read","x")=1032192
> ^dataasalongvariablename45678901(8,"read","x","ref")=1048575

The 8 indicates the 8th iteration (whose message length is set in socket/inref/socset.m)
in the second phase of the test (second invocation of "mumps -run socdev" in
socket/u_inref/socdevice.csh) where a record of length 2**20-1 (i.e. 1048575) gets sent
from the client (client^socdev) through a socket device and the server (server^socdev) reads it.

The test failure symptom indicates that the server read less than the expected length of data.
The actual length was 1032192 bytes whereas we expected to read 1048575 bytes.

The actual length varied in different runs and was not the same. In fact, the failure happened
randomly but always on a system where TLS 1.3 was installed and when gtm_test_tls env var was
set to TRUE by the test framework.

Turns out the client was not waiting for the server read to complete before terminating. This meant
the read() system call on the server side could return a ECONNRESET because the client had terminated
even though data that the client had sent before has not yet been read by the server. I think this is
a TLS 1.3 thing in that it forwards the connection reset message as a higher priority message to the
application ahead of buffered data that has already been read as part of the same connection.

To fix this, the client logic now waits for the server to finish all pending reads before it exits.
This is anyways a good thing to do.

While at this, noticed an unnecessary sleep in socket/u_inref/socdevice.csh. Removed it.

Also noticed a lot of "h 10" and "h 5" usages in the socket test all over. Seems like it is an unnecessary
thing as it means every stage of the test (there are around 25 + 8 stages altogether in the socket/socdevice
subtest) takes 10 seconds to move on causing a total test runtime of about 6 minutes when actually all of
it can be done in a few seconds. But I am not sure if that is there for test coverage or something else
so I am not meddling with that now.