Fix suppl_inst_B/supplementary_err subtest failure due to env_supplementary.csh not being sourced for $MSR STARTSRC
This subtest was fixed yesterday to address a random failure. The fix involved setting the env var
gtm_test_repl_skipsrcchkhlth in the file env_supplementary.csh before $MSR STARTSRC and $MSR STARTRCV
commands. While the STARTRCV worked because it invokes remote_getenv.csh which in turn sources
the env_supplementary.csh, it turns out the STARTSRC does not source env_supplementary.csh. Therefore
the prior commit did not work for the REPLINSTSECLEN case where a $MSR STARTSRC INST4 INST5 was done
(the source side is INST4 which is not local) in the sense the env var setting did not take effect.
Therefore a source server checkhealth happened which caused one more REPLINSTSECLEN to show up in
the output causing two REPLINSTSECLEN messages and a test failure.

The $MSR framework is now enhanced to source env_supplementary.csh for STARTSRC command. This is the
necessary change to fix the test failure.

While at this, the STARTRCV command was also enhanced to do the same to be consistent (even though it
is not necessary because the $sec_getenv call done automatically does a call to env_supplementary.csh).
And com/multisite_replic.awk was enhanced to check gtm_test_repl_skipsrcchkhlth just like it already
does gtm_test_repl_skiprcvrchkhlth because it is possible the source side is on a remote host in
a multi-host replication test. Both these changes are not necessary for the test failure but generally
good framework changes to have.