Enhance v63000a/gtm8522 subtest to record randomly generated pattern strings and atoms in case of a rare test hang
We had a test hang once in this subtest (across hundreds of test runs). Since it was a pro build with no
symbol table information, it was not easy to find out exactly what pattern strings and atoms
(randomly generated by gtm8522b test) was causing the hang. The test now records this information so
we have the actual pattern string and atom accessible in such cases.