Fix rare r120/ydbdist subtest failure due to extra SECSHRNOYDBDIST messages
We had a failure of this test due to multiple SECSHRNOYDBDIST messages showing up in syslog when we
expected just one. This is because prior stages of this test generate similar messages. It is rarely
possible (in a loaded system) for those prior messages to show up right when we are expecting the
last stage message so we could end up with two such messages instead of the expected one message.

To fix this failure, we note down the pid of the gtmsecshr started in the last stage of the test and
search for SECSHRNOYDBDIST messages from that pid in the last stage.