Enhance unicode_io/unicode_fifo subtest to better analyze rare timing related subtest failures
We had a failure of this subtest with the following diff.

> FIFO reader unicodefifo3^unicodefifo couldn't start, exiting

The unicode_fifo subtest starts fifo writer and reader processes separately using JOB commands
and uses global variables with hangs for synchronization. I suspect this is a timing issue in
the test that caused the above failure but exactly what is not sure.

Towards identifying that, the test is now enhanced to do a "do ^sstepgbl" at the beginning.
This records the activities of both writer and reader processes in the same global ^%sstepgbl
with timestamps that should give us a better understanding of the actual interleaved order
of execution of the multiple processes (which should identify the test timing issue).

This is added to an existing line (instead of creating a new line) as otherwise it disturbs
the line #s in the reference file.