Fix tls/errors subtest failure on TLS 1.3 systems due to YDB-W-TLSIOERROR in source server log
After the latest round of fixes to this test, it still fails with the following diff.

171a172,173
> HOST:tls_1_1/errors/SRC_08_46_42_30.log
> %YDB-W-TLSIOERROR, Error during TLS/SSL recv operation

The prior fix to this test is incomplete in that if a TLSIOERROR shows up in the source server
log (which the test expects), it should remove that from the source server log to avoid the
test framework from catching it later and causing a test failure.

This round of fix addresses that.