Fix rare v43001b/init_gld subtest failure on slow systems due to test framework choice of spanning regions
On a slow system, we had a test failure with the following diff

396a397,398
> TEST-E-TIMEOUT, waited for 180 secs, but at least one of the regions were not updated
>  ^a at start:15994 now:60282 ^b at start:12004 now:12004

In this case, some nodes of ^b were mapped to AREG (because of the test framework choice of
gtm_test_spanreg=1). This meant that even though lots of updates to nodes in ^b were happening
(which is what v43001b/inref/wait.m was waiting for above), they were not happening in BREG so
the SET gvstats of BREG were untouched for 3 minutes and so the test timed out.

On a fast system, updates to ^b nodes would eventually go past the range mapping to AREG and
soon get into the range mapping to BREG before the 3 minute timeout so we never saw this failure
till now.

In any case, since wait.m clearly expects a* to map to AREG and b* to map to BREG, it is best to
keep it that way across all systems (slow or fast). That is now done. Should avoid such spanning-regions
related test failures.