[#205] Migrate all subtests in simplapi test to simplethreadapi test (8 subtests still fail and will be fixed in future commits)
A new simplethreadapi test is created based off the existing simpleapi test.

The main type of changes were to convert _s() calls to _st() calls. For example, calls to ydb_get_s()
in the simpleapi test were converted to use ydb_get_st() etc.
There were a lot of other types of changes done. The full list of the changes can be found by
doing a diff of the simpleapi and simplethreadapi test version of a .csh or .c or .m or .txt file
(for example, diff simpleapi/u_inref/gvnget.csh simplethreadapi/u_inref/gvnget.csh).

During the migration, some issues/typos (e.g. tp_get_s -> ydb_get_s) in the existing simpleapi test
were fixed.

And 5 subtests which were implemented in the simpleapi test but had not been enabled were now enabled.
These are the isvsubsnext, isvsubsprev, isvnodenext, isvdata and isvincr subtest. Most of them are
a test of the UNIMPLOP error but it was noticed this error message did not have (SimpleAPI) in the
M-entryref for ydb_zstatus() return value so the code was fixed to address this and the reference files
were fixed too.