Fix go/unit_tests subtest failure on slow systems if they take >= 10 seconds to run
The regular expression .....s accounted for times less than 10 seconds (e.g. 9.845s).
But a slow/loaded system could sometimes take 10 seconds or higher to run this test
and we do not want a failure in that case as long as "ok" is seen. So the regular
expression is now fixed to allow for an arbitrary runtime (i.e. 10.845s or 100.845s etc.)