[#205] Migrate gvnlvnnodenext, gvnlvnnodeprev and isvnodeprev subtests from simpleapi test to simplethreadapi test
Took this opportunity to also reorder the commented section describing the list of subtests
in simpleapi/instream.csh and simplethreadapi/instream.csh so they are in sync and marked
the to-be-implemented subtests with a [....] owner name.

Also undid a change that went in simpleapi/inref/gvnsubsnext3_errors.c as that was incomplete.
Will ensure the complete change happens as a separate commit.