Fix rare/random simpleapi/gvnlvnnodenext and simplethreadapi/gvnlvnnodenext subtest failure
When testing for YDB_ERR_NODEEND, sub is randomly set so it can be checked afterwards to
ensure it does not change after calling ydb_node_next/ydb_node_previous. In some instances
the value of &sub could be set to 0. When this occurs, because there was no reset for the
next part of the test, testing with two branches of nodes, the test would result in a
YDB_ERR_INSUFFSUBS error. This is fixed by setting sub to 1 before this part of the test.