[#205] Fix r124/ydb383 subtest to not use potentially garbage <errstr>
gvnset() is the callback function invoked outside of TP and inside of TP.
But the callback function interface is different for the two cases (ydb_tpfnptr_t vs ydb_tp2fnptr_t).
That would mean duplicating the gvnset() code in two separate functions.

To avoid this duplication, ydb383.c invokes gvnset() with the ydb_tp2fnptr_t prototype.
But that means the "tptoken" and "errstr" values that get passed into gvnset() when the call
is made outside of TP are not reliable. In one test run on a 32-bit ARMV7L box, we saw the
"errstr" value of 0x1 which is a garbage non-null value that when used in the call to "ydb_set_st"
resulted in a SIG-11.

This is now fixed by always passing in NULL instead of in ydb_set_st() in the <errstr> parameter.