Fix errors/test_fao, v62001/hash and v62002/zwritesvn subtest failures on ARMV7L due to not using same compilation flags at link time too (required by LTO)
These are the only subtests that currently use LTO for linking since they link against libmumps.a
instead of libyottadb.so. But they were not using the same flags used at compile time when doing
the LTO link. This works fine on all platforms except 32-bit ARMV7L (e.g. due to specific compiler
flags like -march=armv7-a) where it mattered. And the link failed. This is now fixed by using
the same flags in compile and link on all platforms.